Finsh Primary Task of Features

+++ server/controllers/post-controller/PostController.js
@@ -3,6 +3,7 @@ import { UserPostImageMover } from "../../helper/FileMovers/UserFileMover.js";
 import UserPosts from "../../models/posts/Posts.js";
 import { deleteFile } from "../../utils/FileOperations.js";
 import { getDataFromToken } from "../../utils/getDataFromToken.js";
+import UserTaskSubmission from "../../models/user-academics/UserTaskSubmission.js";
 
 const maxImages = 5;
 export const createPost = async (req, res) => {
@@ -65,10 +66,20 @@ export const deletePost = async (req, res) => {
     const delPost = await UserPosts.findOneAndDelete({ _id: objectId });
     if (!delPost) return res.status(404).json({ error: "Post Not Found" });
 
-    delPost?.images?.map(async (item) => {
-      await deleteFile(`../media/${item?.original}`);
-      await deleteFile(`../media/${item?.compressed}`);
-    });
+    if (delPost?.post_type !== "task") {
+      delPost?.images?.map(async (item) => {
+        await deleteFile(`../media/${item?.original}`);
+        await deleteFile(`../media/${item?.compressed}`);
+      });
+    } else if (delPost?.post_type === "task") {
+      await UserTaskSubmission.findOneAndUpdate(
+        {
+          _id: delPost?.submission_id,
+          status: "approved",
+        },
+        { $set: { is_posted: false } },
+      );
+    }
     return res.status(200).json({ message: "Post Deleted Successfully" });
   } catch (error) {
     console.error(error);
@@ -184,6 +195,114 @@ export const getPostByPostId = async (req, res) => {
     const { post_id } = req.params;
     const posts = await UserPosts.findOne({ _id: post_id });
     return res.status(200).json(posts);
+  } catch (error) {
+    console.error(error);
+    return res
+      .status(500)
+      .json({ error: "Something went wrong. Please try again." });
+  }
+};
+
+export const PostByUserTaskSubmission = async (req, res) => {
+  try {
+    const { submission_id, message, post_type } = req.body;
+    const userId = await getDataFromToken(req);
+
+    const submission = await UserTaskSubmission.findOne({
+      _id: submission_id,
+      status: "approved",
+    });
+    if (!submission) {
+      return res.status(400).json({ error: "Incorrect Submission." });
+    }
+    const images = submission.images;
+
+    const newpost = await UserPosts({
+      userId,
+      images,
+      text: message,
+      post_type,
+      submission_id,
+    });
+    const savedPost = await newpost.save();
+    if (!savedPost)
+      return res
+        .status(500)
+        .json({ error: "Something went wrong. Please try again." });
+
+    await UserTaskSubmission.findOneAndUpdate(
+      {
+        _id: submission_id,
+        status: "approved",
+      },
+      { $set: { is_posted: true } },
+    );
+
+    return res.status(200).json({ message: "Your Activity Post Sucessfully" });
+  } catch (error) {
+    console.error(error);
+    return res
+      .status(500)
+      .json({ error: "Something went wrong. Please try again." });
+  }
+};
+
+export const togglePostPrivate = async (req, res) => {
+  try {
+    const { post_id } = req.params;
+    const authUserId = await getDataFromToken(req);
+
+    if (!mongoose.Types.ObjectId.isValid(post_id)) {
+      return res.status(400).json({ error: "Invalid post id" });
+    }
+
+    const post = await UserPosts.findById(post_id);
+
+    if (!post) {
+      return res.status(404).json({ error: "Post not found" });
+    }
+
+    if (post?.userId?.toString() !== authUserId?.toString()) {
+      return res.status(403).json({ error: "Unauthorized action" });
+    }
+
+    post.is_private = !post.is_private;
+
+    await post.save();
+
+    return res
+      .status(200)
+      .json({ message: "Post privacy updated successfully" });
+  } catch (error) {
+    console.error(error);
+    return res
+      .status(500)
+      .json({ error: "Something went wrong. Please try again." });
+  }
+};
+
+export const updatePostStatus = async (req, res) => {
+  try {
+    const { post_id, status } = req.body;
+    const authUserId = await getDataFromToken(req);
+
+    if (!mongoose.Types.ObjectId.isValid(post_id)) {
+      return res.status(400).json({ error: "Invalid post id" });
+    }
+    if (!["approved", "rejected"]?.includes(status))
+      return res.status(400).json({ error: "Invalid Action" });
+
+    const post = await UserPosts.findById(post_id);
+
+    if (!post) {
+      return res.status(404).json({ error: "Post not found" });
+    }
+
+    post.status = status;
+
+    await post.save();
+
+    return res.status(200).json({ message: "Post updated successfully" });
   } catch (error) {
     console.error(error);
     return res
